arm64: dts: add support for the Pixel 2 XL
Add a minimal devicetree capable of booting on the Pixel 2 XL and Pixel
2 MSM8998 devices. Initial support can boot the device into a Linux
rootfs and expose an ethernet gadget to the host.

The device requires some quirks to boot, notably the dtbo partition used
by Android contains multiple overlays, the bootloader expects an overlay
with a specific ID to be there, most likely to support different
revisions of the hardware.
As a result an empty overlay must be built and flashed to persuade the
bootloader to continue booting, the stock dtb overlays break mainline.

More details and instructions can be found here:
	https://gitlab.com/calebccff/dtbo-google-wahoo-mainline