arm64: dts: qcom: add support for the Pixel 2 XL
Add a minimal devicetree capable of booting on the Pixel 2 XL MSM8998
device.

It's currently possible to boot the device into postmarketOS with USB
networking, however the display panel depends on Display Stream
Compression which is not yet supported in the kernel.

The bootloader also requires that the dtbo partition contains a device
tree overlay with a particular id which has to be overlayed onto the
existing dtb. It's possible to use a specially crafted dtbo partition to
workaround this, more information is available here:

    https://gitlab.com/calebccff/dtbo-google-wahoo-mainline

Signed-off-by: Caleb Connolly <caleb@connolly.tech>
---
It's possible to get wifi working by running Bjorns diag-router in the
background, without this the wifi firmware crashes every 10 seconds or
so. This is the same issue encountered on the OnePlus 5.